move files, create components
