+++ components/pages/file-entity/index.stories.tsx
